code and test improvements
