+++ projects/pt1/python/torch_mlir_e2e_test/test_suite/reduction.py
@@ -755,6 +755,171 @@ def ReduceMinUnsignedIntModule_basic(module, tu: TestUtils):
     module.forward(tu.randint(3, 4, 5, high=100))
 
 # ==============================================================================
+
+class ArgminModule(torch.nn.Module):
+    def __init__(self):
+        super().__init__()
+
+    @export
+    @annotate_args([
+        None,
+        ([-1, -1], torch.float32, True),
+    ])
+
+    def forward(self, a):
+        return torch.ops.aten.argmin(a)
+
+
+@register_test_case(module_factory=lambda: ArgminModule())
+def ArgminModule_basic(module, tu: TestUtils):
+    module.forward(tu.rand(3, 4))
+
+# ==============================================================================
+
+class ArgminIntModule(torch.nn.Module):
+    def __init__(self):
+        super().__init__()
+
+    @export
+    @annotate_args([
+        None,
+        ([-1, -1], torch.int64, True),
+    ])
+
+    def forward(self, a):
+        return torch.ops.aten.argmin(a)
+
+
+@register_test_case(module_factory=lambda: ArgminIntModule())
+def ArgminIntModule_basic(module, tu: TestUtils):
+    module.forward(tu.randint(3, 4, low=-100, high=100))
+
+@register_test_case(module_factory=lambda: ArgminIntModule())
+def ArgminIntModule_multiple_mins(module, tu: TestUtils):
+    # To cover the special case that the minimal value occurs more than once.
+    # The pytorch convention is here to consider the first occurence as the argmin.
+    module.forward(torch.full((3,4), tu.randint(1).item(), dtype=torch.int64))
+
+# ==============================================================================
+
+class ArgminWithDimModule(torch.nn.Module):
+    def __init__(self):
+        super().__init__()
+
+    @export
+    @annotate_args([
+        None,
+        ([-1, -1, -1], torch.float32, True),
+    ])
+    def forward(self, a):
+        return torch.ops.aten.argmin(a, dim=1)
+
+@register_test_case(module_factory=lambda: ArgminWithDimModule())
+def ArgminModule_with_dim(module, tu: TestUtils):
+    module.forward(tu.rand(3, 4, 5))
+
+# ==============================================================================
+
+class ArgminKeepDimsModule(torch.nn.Module):
+    def __init__(self):
+        super().__init__()
+
+    @export
+    @annotate_args([
+        None, 
+        ([-1, -1], torch.float32, True),
+    ])
+    def forward(self, a):
+        return torch.ops.aten.argmin(a, 0, True)
+
+@register_test_case(module_factory=lambda: ArgminKeepDimsModule())
+def ArgminModule_keepDim(module, tu: TestUtils):
+    module.forward(tu.rand(4, 6))
+
+# ==============================================================================
+
+class ArgmaxModule(torch.nn.Module):
+    def __init__(self):
+        super().__init__()
+
+    @export
+    @annotate_args([
+        None,
+        ([-1, -1], torch.float32, True),
+    ])
+
+    def forward(self, a):
+        return torch.ops.aten.argmax(a)
+
+
+@register_test_case(module_factory=lambda: ArgmaxModule())
+def ArgmaxModule_basic(module, tu: TestUtils):
+    module.forward(tu.rand(3, 4))
+
+# ==============================================================================
+
+class ArgmaxIntModule(torch.nn.Module):
+    def __init__(self):
+        super().__init__()
+
+    @export
+    @annotate_args([
+        None,
+        ([-1, -1], torch.int64, True),
+    ])
+
+    def forward(self, a):
+        return torch.ops.aten.argmax(a)
+
+
+@register_test_case(module_factory=lambda: ArgmaxIntModule())
+def ArgmaxIntModule_basic(module, tu: TestUtils):
+    module.forward(tu.randint(3, 4, low=-100, high=100))
+
+@register_test_case(module_factory=lambda: ArgmaxIntModule())
+def ArgmaxIntModule_multiple_maxs(module, tu: TestUtils):
+    # To cover the special case that the maximal value occurs more than once.
+    # The pytorch convention is here to consider the first occurence as the argmax.
+    module.forward(torch.full((3,4), tu.randint(1).item(), dtype=torch.int64))
+
+# ==============================================================================
+
+class ArgmaxWithDimModule(torch.nn.Module):
+    def __init__(self):
+        super().__init__()
+
+    @export
+    @annotate_args([
+        None,
+        ([-1, -1, -1], torch.float32, True),
+    ])
+    def forward(self, a):
+        return torch.ops.aten.argmax(a, dim=1)
+
+@register_test_case(module_factory=lambda: ArgmaxWithDimModule())
+def ArgmaxModule_with_dim(module, tu: TestUtils):
+    module.forward(tu.rand(3, 4, 5))
+
+# ==============================================================================
+
+class ArgmaxKeepDimsModule(torch.nn.Module):
+    def __init__(self):
+        super().__init__()
+
+    @export
+    @annotate_args([
+        None, 
+        ([-1, -1], torch.float32, True),
+    ])
+    def forward(self, a):
+        return torch.ops.aten.argmax(a, 0, True)
+
+@register_test_case(module_factory=lambda: ArgmaxKeepDimsModule())
+def ArgmaxModule_keepDim(module, tu: TestUtils):
+    module.forward(tu.rand(4, 6))
+
+# ==============================================================================
+
 class ReduceL1NormModule(torch.nn.Module):
     def __init__(self):
         super().__init__()
